fix: green the workspace — 2452 tests, clippy clean, fmt clean
Two guards were failing and both were stale rather than broken code:

- w4a16_exact_lm_head_routing_static forbade w4a16_gemm_n64_m32_ldb outright.
  That route is now a deliberate tensor-core re-reference behind the
  ATLAS_LM_HEAD_TC gate (REFREEZE 2026-08-17, hash 12e0c0ad). The guard now
  asserts the gate rather than the absence of the call, so an UNGATED
  appearance still fails. The other three lossy routes remain forbidden
  outright and are verified absent.

- swap_lifecycle_source_scan pins scheduler/mod.rs by exact byte length and
  FNV-1a hash. Unrelated module and import additions tripped it. Verified the
  swap-out region is byte-identical (victim selection still filters
  grammar_state.is_none() and the let-else warns when all sequences are
  grammar-active) before re-blessing the constants.

Also clears 19 clippy lints (clippy::all is deny-level workspace-wide). The
three unwrap-after-is_some sites became Option::filter so the
prefill_proj_fast_enabled gate is preserved without an unwrap. Applies rustfmt
to 10 files and drops a tracked async_propose.rs.orig merge artifact.

+++ crates/spark-model/src/layers/dflash_head/forward_block_layer.rs
@@ -216,7 +216,7 @@ impl BlockDiffusionDraftHead {
     ) -> Result<()> {
         use crate::layers::ops;
         let gpu = ctx.gpu;
-        let noise_bytes = noise_count as usize * conv.groups as usize * 16 * 2;
+        let noise_bytes = noise_count as usize * conv.groups * 16 * 2;
         ops::dflash2_conv_finish(
             gpu,
             self.kernels.dflash2_conv_finish,
